+++ src/events/mod.rs
@@ -1,4 +1,3 @@
-
 mod event;
 mod event_source;