+++ src/text.rs
@@ -2,13 +2,20 @@ use std::fmt;
 
 use minimad::Text;
 
-use crate::skin::MadSkin;
 use crate::code;
 use crate::line::FmtLine;
+use crate::skin::MadSkin;
 use crate::tbl;
 use crate::wrap;
 
 /// a formatted text, implementing Display
+/// ```
+/// use termimad::*;
+/// let skin = MadSkin::default();
+/// let my_markdown = "#title\n* item 1\n* item 2";
+/// let text = FmtText::from(&skin, &my_markdown, Some(80));
+/// println!("{}", &text);
+/// ```
 pub struct FmtText<'k, 's> {
     pub skin: &'k MadSkin,
     pub lines: Vec<FmtLine<'s>>,
@@ -16,11 +23,17 @@ pub struct FmtText<'k, 's> {
 }
 
 impl<'k, 's> FmtText<'k, 's> {
+    /// build a displayable text for the specified width and skin
+    ///
+    /// This can be called directly or using one of the skin helper
+    /// method.
     pub fn from(skin: &'k MadSkin, src: &'s str, width: Option<usize>) -> FmtText<'k, 's> {
         let mut mt = Text::from(src);
-        let mut lines = mt.lines.drain(..).map(
-            |mline| FmtLine::from(mline, skin)
-        ).collect();
+        let mut lines = mt
+            .lines
+            .drain(..)
+            .map(|mline| FmtLine::from(mline, skin))
+            .collect();
 
         tbl::fix_all_tables(&mut lines, width.unwrap_or(std::usize::MAX));
         code::justify_blocks(&mut lines);
@@ -28,11 +41,7 @@ impl<'k, 's> FmtText<'k, 's> {
             lines = wrap::hard_wrap_lines(lines, width);
         }
 
-        FmtText {
-            skin,
-            lines,
-            width,
-        }
+        FmtText { skin, lines, width }
     }
 }
 
@@ -45,4 +54,3 @@ impl fmt::Display for FmtText<'_, '_> {
         Ok(())
     }
 }
-